paychain-examples/requirements.txt
python-dotenv>=0.19.0
langchain>=0.1.0
langchain-openai>=0.0.3
langchain-community>=0.0.10
langchain-core>=0.1.0
openai>=1.1.0
langchain-payman-tool>=0.1.0 